+++ src/windows/BBX_PopupPanel.c
@@ -101,7 +101,7 @@ static LRESULT CALLBACK WndProc(HWND hwnd, UINT msg, WPARAM wParam, LPARAM lPara
   case WM_CREATE:
 	  pan = (BBX_Panel *)((CREATESTRUCT *)lParam)->lpCreateParams;
 	  pan->win = hwnd;
-	  SetWindowLongPtr(hwnd, GWLP_USERDATA, (LONG)((CREATESTRUCT *)lParam)->lpCreateParams);
+	  SetWindowLongPtr(hwnd, GWLP_USERDATA, ((CREATESTRUCT *)lParam)->lpCreateParams);
 	  break;
   case WM_SIZE:
 	  /*
